Temporarily workaround transcripts being incorrect on edited messages.
There seems to be wrong with the edit hook; my assumption is that it
doesn't update the linked message after the edit, leading to stale
embeds being served. I tried debugging it, but could not figure out what
was the issue; this workaround is thus temporary until I get around to
fixing it properly.

Note that this requires an extra round-trip to Discord, so the “Show
Transcripts” button takes a little bit longer than it did previously.
It's not too long to be annoying, however.