Remove unnecessary emoji variation selectors.
Note: throughout this commit message, all hex sequences are for text
encoded in UTF-8, which is the character encoding this commit message
should have be encoded as.

There are multiple ways to represent the same code-point in Unicode, and
these variations are controlled with “Variation Selectors” (there are 16
of these, and they have their own block). The relevant ones here are:
  - just the base code-point (no variation selector): 👍 (f09f 918d);
  - the text variation selector (VS-15)¹: 👍︎ (f09f 918d efb8 8e²); and
  - the emoji variation selector (VS-16): 👍️ (f09f 918d efb8 8f³).

If you use Discord, you likely know that all emojis are converted to
SVGs that use Twemoji and are clickable⁴. However, for whatever reason,
Discord seems to *only* convert emojis⁵ that are just the base
code-point to SVGs. This makes sense in the case of the text variation,
where it likely looks different. However, this can be an annoyance with
emojis that have the emoji variation selector, since they use the system
emoji font and aren't interactive, but other applications wouldn't show
any difference between these two in most cases. These variation
selectors snuck in over time whenever I inserted emojis, since the emoji
picker I use⁶ includes the emoji variation selector for some emojis that
don't strictly need it (but not all, for some reason), which results in
them not being converted to SVGs by Discord.

This commit hence strips the emoji variation selector from most (see
below) emojis in the code.

There is an exception though: while most emojis default to the emoji
presentation, not all do; an example is the heart emoji:
  - base code-point (no variation selector): ❤ (e29d a4)
  - text variation selector (VS-15): ❤︎ (e29d a4ef b88e²)
  - emoji variation selector (VS-16): ❤️ (e29d a4ef b88f³)

Assuming a complaint application, both the first and the second will
likely be black or white, and single-width rather than double-width;
while the second would be the heart emoji from your emoji font⁷, and
double-width. If this is not what you see, try a browser⁸.

Luckily, Discord handles these correctly, and handles these with the
emoji variation when using its emoji picker, and also converts emojis
that have the emoji variation selector for these emojis to SVGs. The
base code-point remains as the text variation for these, and the text
variation also remains as the text variation; a literal emoji character
that is presented with the emoji variation cannot be acquirable without
escaping it⁵. This means that the emoji variation selector must remain,
for a different variation would be used were it to be removed; this is
the reason why not all variation selectors have been removed.

¹Some applications may get confused or break because of this one; paste
 it into a browser⁸ if you're having issues viewing them in your
 favourite terminal emulator or text editor, for example. It may also
 show up as the emoji variation rather than the intended text variation,
 depending on your system configuration⁹.

²The Unicode code-point of VS-15 is U+FE0E.

³The Unicode code-point of VS-16 is U+FE0F.

⁴You may have found that Discord (since at least 2024) also makes the
 SVGs slightly larger than the line height, which makes actual emoji
 characters look slightly smaller as well.

⁵A backslash before an emoji always escapes it, which makes Discord
 avoid touching it. This means that a backslash can make an emoji be
 forced to use the system emoji font, rather than being converted to
 a SVG.

⁶The emoji picker I use is Smile: https://smile.mijorus.it.

⁷Unless you changed it, the default is Noto Sans Emoji on Android and
 most Linux distros (and probably BSDs too), Apple Color Emoji on iOS,
 iPadOS, macOS, visionOS, watchOS, and other Apple operating systems,
 and Segoe UI on Windows.

⁸Below is some sample HTML you can save in a file to load up in your
 browser, so that you can paste this commit message inside the text box
 for viewing emojis:

    <!DOCTYPE html>
    <html>
        <body>
            <textarea rows="99" cols="72"></textarea>
        </body>
    </html>

⁹On my system, all chromium-based browsers and hence electron
 applications, GTK applications, Qt applications, and some other
 assorted ones, seem to use the emoji variation for all three, despite
 my system having a text variation available in the Deja Vu font. I'm
 not sure why.  Firefox seems to display them “correctly”, though, at
 least for me. This doesn't seem to happen with the heart emoji later
 on; I'm not very familiar with how the variation selectors are
 rendered, so there may be some underlying reason for this. It may also
 be because of weirdness with Fontconfig and the hacks employed by the
 ttf-twemoji AUR package
 (https://aur.archlinux.org/cgit/aur.git/tree/75-twemoji.conf?h=ttf-twemoji&id=6703f149f607db59fd4c86efcfad923342d5a985#n52),
 which is what I use to set my emoji font to Twemoji globally.