Correctly handle GitHub comments with no reactions.
I did handle it, with `if formatted_reactions:`, and I tested it too.
Then in a review comment [1], trag1c suggested a more concise way to
write the formatted_reactions comprehension, and so I did it, by
deleting the ternary and the square brackets, dedenting  the contents of
the comprehension (or maybe I saved the file and let `ruff format`
handle it; I no longer remember), and leaving the brackets that were
originally around the ternary. Which is cool and all until you realise
that the suggestion had this code:

    formatted_reactions = comment.reactions and [
        f"{REACTION_EMOJIS[reaction]} ×{count}"  # noqa: RUF001
        for reaction, count in comment.reactions
        if count
    ]

And I ended up [2] with this code:

    formatted_reactions = comment.reactions and (
        f"{REACTION_EMOJIS[reaction]} ×{count}"  # noqa: RUF001
        for reaction, count in comment.reactions
        if count
    )

Which is subtly different because my code is lazy. Which of course works
just fine when I tested a random comment that had reactions, since it is
consumed just fine to give the same output. But generators are truthy.
Which made `if formatted_reactions:` always run, leaving a `-#` all
alone by itself in an embed field.  And I didn't bother test a comment
without reactions too, so I didn't catch it until the next time
I actually linked a comment.

[1]: https://github.com/ghostty-org/discord-bot/pull/241#discussion_r2156587834
[2]: https://github.com/ghostty-org/discord-bot/compare/bdf97549c988a78c3d48bd493026868b162869e9..8374794951c371ee996ee77fe771388d596af862#diff-e430f38e0a5c258e4342f80c54469530982497aa9c5e22f6f028ef2c90bd47ddR49-R53