Set httpx's log level to WARNING.
The only logs in httpx's code is to print out every REST request at
level INFO. This disables those since they're rarely going to be useful
for the developer. Since this is the only logging they currently have as
of 0.28.1, this effectively disables logs from httpx; it's done this way
instead of disabling the httpx library logs entirely in case they add
more logging in the future.