Reply instead of DM caller in !sync.
We exclusively run it in private channels anyway, and were it to be run
outside (which hasn't happened yet), it would almost certainly be for
demonstration purposes, where the sync message being a reply would be
a better outcome.

The command ignores the case where a non-mod runs !sync, so all other
users would not notice anything.